# replace old libsecp256k1 fork with upstream; boosts k1 recovery performance