--- a/libraries/chain/include/eosio/chain/types.hpp
+++ b/libraries/chain/include/eosio/chain/types.hpp
@@ -74,9 +74,6 @@ namespace eosio { namespace chain {
    using                               fc::flat_multimap;
    using                               fc::flat_set;
    using                               std::variant;
-   using                               fc::ecc::range_proof_type;
-   using                               fc::ecc::range_proof_info;
-   using                               fc::ecc::commitment_type;
 
    using public_key_type  = fc::crypto::public_key;
    using private_key_type = fc::crypto::private_key;