@spoonincode 2021-01-27
Hey guys, we really need to approve and merge EOSIO/fc#180 first. Otherwise, now develop's submodule is no longer pointing to fc's eosio branch